refactored user entity and related classes to use enums for role and status; implemented ApiResponse for consistent response structure

+++ src/main/java/com/cts/elearn/mapper/UserResponseMapper.java
@@ -9,7 +9,7 @@ public class UserResponseMapper {
     // MAPPER
     public UserResponse mapToResponse(User user) {
         return new UserResponse(
-                user.getId().intValue(),
+                user.getId(),
                 user.getName(),
                 user.getContactNumber(),
                 user.getEmail()
